+++ openyich-framework/src/main/java/com/openyich/framework/boot/aspectj/lang/StopWatch.java
@@ -1,0 +1,14 @@
+package com.openyich.framework.boot.aspectj.lang;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
+public @interface StopWatch {
+
+  String value() default "";
+  
+}