+++ openyich-framework/src/main/java/com/openyich/framework/boot/errors/AlreadyUsedException.java
@@ -1,33 +1,44 @@
 package com.openyich.framework.boot.errors;
 
+import static org.apiguardian.api.API.Status.STABLE;
+
 import java.util.Map;
 
+import org.apiguardian.api.API;
 import org.zalando.problem.Status;
-import org.zalando.problem.ThrowableProblem;
 
+@API(status = STABLE)
 public class AlreadyUsedException extends OpenYichThrowableProblem {
 
   private static final long serialVersionUID = 1L;
 
-  public AlreadyUsedException(String message) {
-    super("Already Used", Status.CONFLICT, message);
+  public AlreadyUsedException(String code, String message, Map<String, Object> parameters) {
+    super(code, Status.CONFLICT, message, parameters);
   }
 
-  public AlreadyUsedException(String message, Map<String, Object> parameters) {
-    super("Already Used", Status.CONFLICT, message, parameters);
+  public AlreadyUsedException(String code, String message, String... params) {
+    super(code, Status.CONFLICT, message, params);
   }
 
-  public AlreadyUsedException(String message, String... params) {
-    super("Already Used", Status.CONFLICT, message, params);
-  }
-
-  public AlreadyUsedException(String message, ThrowableProblem cause,
+  public AlreadyUsedException(String code, String message, Throwable cause,
       Map<String, Object> parameters) {
-    super("Already Used", Status.CONFLICT, message, cause, parameters);
+    super(code, Status.CONFLICT, message, cause, parameters);
   }
 
-  public AlreadyUsedException(String message, ThrowableProblem cause, String... params) {
-    super("Already Used", Status.CONFLICT, message, cause, params);
+  public AlreadyUsedException(String code, String message, Throwable cause, String... params) {
+    super(code, Status.CONFLICT, message, cause, params);
+  }
+
+  public AlreadyUsedException(String code, String message) {
+    super(code, Status.CONFLICT, message);
+  }
+
+  public AlreadyUsedException(String code, Throwable cause) {
+    super(code, Status.CONFLICT, cause);
+  }
+
+  public AlreadyUsedException(String code) {
+    super(code, Status.CONFLICT);
   }
 
 }