+++ openyich-framework/src/main/java/com/openyich/framework/boot/errors/BadRequestException.java
@@ -8,36 +8,36 @@ import org.apiguardian.api.API;
 import org.zalando.problem.Status;
 
 @API(status = STABLE)
-public class ValidationException extends OpenYichThrowableProblem {
+public class BadRequestException extends OpenYichThrowableProblem {
 
   private static final long serialVersionUID = 1L;
 
-  public ValidationException(String code, String message, Map<String, Object> parameters) {
+  public BadRequestException(String code, String message, Map<String, Object> parameters) {
     super(code, Status.BAD_REQUEST, message, parameters);
   }
 
-  public ValidationException(String code, String message, String... params) {
+  public BadRequestException(String code, String message, String... params) {
     super(code, Status.BAD_REQUEST, message, params);
   }
 
-  public ValidationException(String code, String message, Throwable cause,
+  public BadRequestException(String code, String message, Throwable cause,
       Map<String, Object> parameters) {
     super(code, Status.BAD_REQUEST, message, cause, parameters);
   }
 
-  public ValidationException(String code, String message, Throwable cause, String... params) {
+  public BadRequestException(String code, String message, Throwable cause, String... params) {
     super(code, Status.BAD_REQUEST, message, cause, params);
   }
 
-  public ValidationException(String code, String message) {
+  public BadRequestException(String code, String message) {
     super(code, Status.BAD_REQUEST, message);
   }
 
-  public ValidationException(String code, Throwable cause) {
+  public BadRequestException(String code, Throwable cause) {
     super(code, Status.BAD_REQUEST, cause);
   }
 
-  public ValidationException(String code) {
+  public BadRequestException(String code) {
     super(code, Status.BAD_REQUEST);
   }